Fix (prior DashboardPage.test.tsx) CodeInsightsRootPage.test.tsx
